Create a serde folder with high level serializer serde.py and torch serializer torch_serde.py and native python serializer native_serde.py
