+++ syft/serde/__init__.py
@@ -1,0 +1,2 @@
+from syft.serde import serde
+from syft.serde import torch_serde
